Various fixes and features (subpath, parsing, path regex, gradientTransform)
- Fix to allow reading gradientTransform from inherited href link.
Small fix: gradientTransform was ignored when the gradient was using a
reference link to another defined gradient. This fix allows the correct
reading of the gradientTransform also from inherited href links.

- Added support for the transformGradient attribute that was missing
(only Android and iOS for now)

- Added support for the transformGradient attribute that was missing.
Only gradients without transforms were rendered correctly before. Now
on Android and iOS the transformGradient matrix is parsed and applied
to the LinearGradient and to the RadialGradient. Other platforms are
still ToDo.

- Added a better way to split path data numbers. Some svg files were
not parsed correctly due to some strange formatting. I added a better
way to split path data numbers. Now it works with all numbers
formatting. I used this regex from the SVGO project:
@"([-+]?((\d*\.\d+)|(\d+))([eE][-+]?\d+)?)"; You can test it here:
https://regex101.com/r/9Ryu0U/1

- Fix for the SVG SubPath Issue. Now subpaths are rendered correctly.